Tidying up etc.
