+++ frontend/web-app/src/App.js
@@ -1,11 +1,9 @@
 import './App.css';
-import Navbar from './Navbar';
 import Main from './Main';
 
 function App() {
   return (
     <div className="App">
-      <Navbar />
       <Main />
     </div>
   );
