+++ frontend/web-app/src/Developer.css
@@ -1,4 +1,4 @@
-.Developer-body {
+.Developer {
     height: calc(100vh - 60px);
     margin-top: 60px;
     padding-bottom: 20px;
