+++ frontend/web-app/src/Developer.js
@@ -1,11 +1,14 @@
-import DeveloperBody from "./DeveloperBody";
-import Navbar from "./Navbar";
+import './Developer.css';
 
 const Developer = () => {
     return (
         <div className="Developer">
-            <Navbar />
-            <DeveloperBody />
+            <div className='main-container'>
+                <div className='form'>
+                    <h2>Publikuj alerty bezpieczeństwa bezpośrednio na <span className='outer'><span className='inner'>Cyber</span>ly</span> za pomocą naszego API!</h2>
+                    <button>Aplikuj</button>
+                </div>
+            </div>
         </div>
     );
 }
