--- frontend/web-app/src/DeveloperBody.js
@@ -1,16 +1,0 @@
-import './Developer.css';
-
-const DeveloperBody = () => {
-    return (
-        <div className="Developer-body">
-            <div className='main-container'>
-                <div className='form'>
-                    <h2>Publikuj alerty bezpieczeństwa bezpośrednio na <span className='outer'><span className='inner'>Cyber</span>ly</span> za pomocą naszego API!</h2>
-                    <button>Aplikuj</button>
-                </div>
-            </div>
-        </div>
-    );
-}
-
-export default DeveloperBody;
