+++ frontend/web-app/src/Main.css
@@ -12,6 +12,10 @@
     align-items: center;
 }
 
+.Main.loading {
+    height: calc(100vh - 60px);
+}
+
 .filter {
     margin-top: 20px;
 }
