+++ frontend/web-app/src/PageNotFound.css
@@ -1,0 +1,37 @@
+.error {
+    margin-top: 60px;
+    height: calc(100vh - 60px);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    background: #010811bd;
+    animation-name: pulse;
+    backdrop-filter: saturate(1);
+    animation-duration: 3s;
+    animation-iteration-count: infinite;
+}
+
+.not-found-info {
+    font-size: 52px;
+}
+
+.home-link {
+    text-decoration: none;
+    font-size: 24px;
+    color: rgb(0, 102, 255);
+    font-weight: 500;
+    transition: ease-in-out 200ms;
+}
+
+.home-link:hover {
+    transform: scale(1.05);
+    color: rgb(0, 102, 255);
+}
+
+@keyframes pulse {
+    0%   {backdrop-filter: saturate(1);}
+    25%  {backdrop-filter: saturate(1.8);}
+    50%  {backdrop-filter: saturate(1.3);}
+    100% {backdrop-filter: saturate(1);}
+  }
